Predicting client's payoff using ML model
In this project, I am working with financial lending data from Lending Club. Lending Club is a American marketplace for personal loans that matches borrowers who are seeking a loan with investors looking to lend money and make a return.

HERE IS HOW LENDING CLUB WORKS :

Each borrower fills out a comprehensive application, providing their past financial history, the reason for the loan, and more. Lending Club evaluates each borrower's credit score using past historical data and assign an interest rate to the borrower.
A higher interest rate means that the borrower is riskier and more unlikely to pay back the loan. In contrast, a lower interest rate means that the borrower has a good credit history and is more likely to pay back the loan. The interest rates range from 5.32% all the way to 30.99%. If the borrowers accepts the interest rate, then the loan is listed on the Lending CLub marketplace.
Investor are primarily interested in receiving a return on their investments. Approved loans are lsited on the Lending Club website, where qualified investors can browse the borrower's information. Once they're ready to back a loan, the select the amount of money they want to fund. Ultimately, once a loan's requested amount is fully funded, the borrower receives the money minus the origination fee charged by Lending Club.
The borrower then makes monthly payments back to Lending Club either over 36 months or 60 months. Lending Club redistributes these payments to the investors. This means that investors don't have to wait until the full amount is paid off to start to see money back.
In reality, many loans aren't completely paid off on time, however, and some borrowers default on the loan.
Most investors use a portfolio strategy to invest into small amounts in many loans, with healthy mixes of low, medium and interest loans. In this project, however, I'll focus on the mindset of a conservative investor who only wants to invest in the loans that have a good chance of being paid off on time.

Consequently, this is our problematic question:

"Can we build a machine learning model that can accurately predict if a borrower will pay off their loan on time or not ?"

I am going to walkthrough 3 parts: the data cleaning, the feature selection and the prediction making.